Prevent baby animals from spawning
When you right click an animal that has a baby type with their
respective spawn egg it will create a baby animal and remove the spawn
egg.  As this plugin is intended to make eggs spawners, this can cause
players to lose their spawners in exchange for a baby animal.  This
prevents that from happening.